@@ -1,20 +1,20 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="Helpers.fs" />
-    <Compile Include="TestApi.fs" />
-    <Compile Include="*.fs" Exclude="Helpers.fs; TestApi.fs; Program.fs" />
-    <Compile Include="Program.fs" />
+    <Compile Include="Helpers.fs"/>
+    <Compile Include="TestApi.fs"/>
+    <Compile Include="*.fs" Exclude="Helpers.fs; TestApi.fs; Program.fs"/>
+    <Compile Include="Program.fs"/>
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="../Serilog.Sinks.Scalyr/Serilog.Sinks.Scalyr.csproj" />
+    <ProjectReference Include="../Serilog.Sinks.Scalyr/Serilog.Sinks.Scalyr.csproj"/>
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="Expecto" Version="*" />
-    <PackageReference Include="Hornbill" Version="*" />
-    <PackageReference Include="Serilog" Version="*" />
+    <PackageReference Include="Expecto" Version="9.0.4"/>
+    <PackageReference Include="Hornbill" Version="2.6.2"/>
+    <PackageReference Include="Serilog" Version="2.10.0"/>
   </ItemGroup>
 </Project>
