@@ -1,12 +1,11 @@
 using System.Collections.Generic;
 
-namespace Serilog.Sinks.Scalyr
+namespace Serilog.Sinks.Scalyr;
+
+class ScalyrSession
 {
-  internal class ScalyrSession
-  {
     public string Token { get; set; }
     public string Session { get; set; }
     public IEnumerable<ScalyrEvent> Events { get; set; }
     public object SessionInfo { get; set; }
-  }
 }
