@@ -3,5 +3,4 @@
 open Expecto
 
 [<EntryPoint>]
-let main args =
-  runTestsInAssembly defaultConfig args
+let main args = runTestsInAssembly defaultConfig args
